watai-capstone/requirements.txt
langchain>=0.1.0
langchain-mongodb>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
pymongo>=4.6.0
openai>=1.0.0
python-dotenv>=1.0.0
wikipedia>=1.4.0

